pyproject.toml
[tool.poetry]
name = "mosqlimate"
version = "1.1.0"  # changed by semantic-release
description = "TODO"
authors = ["Flávio Codeço Coelho <you@example.com>"]
license = "GNU"
readme = "README.md"


[tool.poetry.dependencies]
python = ">=3.11.0,<=3.12.4"
django-ninja = "^1.2.2"
django-extensions = "^3.2.1"
dr-scaffold = "^2.1.2"
jinja2 = "^3.1.4"
psycopg2-binary = "^2.9.6"
django = "^4.2.2"
python-environ = "^0.4.54"
django-allauth = "^0.54.0"
fontawesomefree = "^6.4.0"
websocket-client = "0.49.0"
uvicorn = "^0.23.2"
gunicorn = "^21.2.0"
whitenoise = "^6.5.0"
pydantic = ">=2.0,<3.0.0"
python-dotenv = "<0.22.0"
msgspec = "^0.18.2"
six = "^1.16.0"
django-plotly-dash = "^2.2.2"
pandas = "^2.1.2"
dpd-static-support = "^0.0.5"
channels = "^4.0.0"
daphne = "^4.0.0"
redis = "^5.0.1"
django-redis = "^5.4.0"
channels-redis = "^4.1.0"
matplotlib = "^3.8.1"
geopandas = "^0.14.0"
pyproj = "^3.6.1"
altair = "^5.1.2"
celery = "^5.3.5"
django-celery-beat = "^2.5.0"
ipython = "^8.18.1"
django-cors-headers = "^4.3.1"
epiweeks = "^2.3.0"
duckdb = "^0.10.0"
dask = {extras = ["dataframe"], version = "^2024.7.0"}
loguru = "^0.7.2"
mosqlient = "^1.9.3"
scipy = "^1.15.2"
scoringrules = "^0.7.1"
scikit-learn = "^1.6.1"
mosqlimate-assistant = { git = "https://github.com/Mosqlimate-project/mosqlimate-assistant.git", rev = "1.7.2" }
watchdog = "^6.0.0"
websockets = "^15.0.1"
django-vectortiles = "^1.0.1"
markdown = "^3.8.2"
markdown2 = {extras = ["all"], version = "^2.5.3"}


[tool.poetry.group.pre-commit.dependencies]
pre-commit = "^3.3.2"
black = "^25.1.0"
ruff = "^0.11.8"


[tool.poetry.group.dev.dependencies]
compose-go = "^2.22.0"
containers-sugar = "1.13.0"
makim = ">=1.15.2"
pygments = "^2.15.1"


[tool.poetry.group.docs.dependencies]
mkdocs = "^1.5.2"
notebook = "^7.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
